fix: 모바일 온디바이스 수어 인식 파이프라인 정렬 및 프리뷰 개선

See merge request s14-final/S14P31A404!166

+++ mobile/app/src/main/java/com/ssafy/mobile/core/vision/inference/SignModelVariant.kt
@@ -15,6 +15,6 @@ enum class SignModelVariant(
     ;
 
     companion object {
-        val DEFAULT = FLOAT32
+        val DEFAULT = FLOAT16
     }
 }
